+++ app/composables/useChartConfig.ts
@@ -1,8 +1,7 @@
 import type { ApexOptions } from "apexcharts";
 
-export function useChartConfig(chartRef: Ref<HTMLElement | null>) {
-  const primaryColor = useCssVar("--ui-primary", chartRef);
-  const textColor = useCssVar("--ui-text", chartRef);
+export function useChartConfig() {
+  // These values can be undefined and it is causing ts errors for ApexOptions since colors can't be undefined there.
 
   const baseOptions = computed<ApexOptions>(() => ({
     chart: {
@@ -42,14 +41,15 @@ export function useChartConfig(chartRef: Ref<HTMLElement | null>) {
     grid: {
       show: false,
     },
+    // I am asserting that useCSSVars is not undefined which I don't think it can be since it should always be defined in the CSS, but I need to do this to satisfy the ApexOptions type.
     dataLabels: {
       enabled: true,
       style: {
         fontFamily: "Noto Sans, sans-serif",
-        colors: [textColor.value],
+        colors: ["var(--ui-text)"],
       },
     },
-    colors: [primaryColor.value],
+    colors: ["var(--ui-primary)"],
   }));
   return {
     baseOptions,