+++ app/pages/changelog.vue
@@ -14,6 +14,31 @@ useSeoMeta({
 
 const versions = [
   {
+    title: "1.7.0",
+    date: "2026-05-18",
+    content: `
+### Features
+
+- Added toggleable background image to the game
+- Added extra information to ability game results to show if the additional minigame was successfully completed
+- Added filter to the archive page to hide completed games
+
+### Changes
+
+- Added Styanax Prime to the game
+- Added Sirius and Orion to the game
+- Updated the style of disabled items in the warframe search to be clearer that they cannot be selected
+
+### Bug Fixes
+
+- Fixed ability archive games before a certain date not properly loading image
+- Fixed global stats lowering or not updating after completing a game
+- Updated grid filtering logic to correctly work to lower the chances of repeat categories or category pairs in the grid game
+- Fixed incorrect data saved for certain warframes that caused them to not show up in the game
+    `,
+    badge: "v1.7.0",
+  },
+  {
     title: "1.6.1",
     date: "2026-04-30",
     content: `